refactor: move packages and add comments
