+++ shared/src/androidMain/kotlin/co/touchlab/kampkit/base/ViewModel.kt
@@ -1,0 +1,13 @@
+package co.touchlab.kampkit.base
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.CoroutineScope
+
+actual abstract class ViewModel actual constructor() : ViewModel() {
+    actual val viewModelScope: CoroutineScope = androidXViewModelScope
+
+    actual override fun onCleared() {
+        super.onCleared()
+    }
+}
