+++ shared/src/commonMain/kotlin/co/touchlab/kampkit/base/StaleDataDelegate.kt
@@ -1,8 +1,12 @@
 package co.touchlab.kampkit.base
 
+/**
+ * Handles updating/checking if data is stale given a [StaleDataKey]
+ */
 interface StaleDataDelegate {
 
     fun updateLastTime(key: StaleDataKey)
 
     fun isDataStale(key: StaleDataKey): Boolean
+
 }
