+++ shared/src/commonMain/kotlin/co/touchlab/kampkit/base/StaleDataKey.kt
@@ -1,5 +1,8 @@
 package co.touchlab.kampkit.base
 
+/**
+ * Enum class that holds a key and length of time for stale data
+ */
 enum class StaleDataKey(val keyName: String, val msUntilStale: Int) {
     BreedStaleData("BreedTimestampKey", ONE_MINUTE_MS * 10);
 }
