+++ shared/src/commonMain/kotlin/co/touchlab/kampkit/base/ViewModel.kt
@@ -1,4 +1,4 @@
-package co.touchlab.kampkit.models
+package co.touchlab.kampkit.base
 
 import kotlinx.coroutines.CoroutineScope
