+++ shared/src/commonMain/kotlin/co/touchlab/kampkit/ktor/Api.kt
@@ -1,9 +1,13 @@
 package co.touchlab.kampkit.ktor
 
 import co.touchlab.kampkit.base.ApiStatus
-import co.touchlab.kampkit.ktor.dto.BreedDto
+import co.touchlab.kampkit.networkmodels.BreedDto
 
+/**
+ * Interface to represent all network calls
+ */
 interface Api {
 
     suspend fun getBreeds(): ApiStatus<BreedDto>
+
 }
