+++ shared/src/iosMain/kotlin/co/touchlab/kampkit/base/ViewModel.kt
@@ -1,0 +1,30 @@
+package co.touchlab.kampkit.base
+
+import kotlinx.coroutines.MainScope
+import kotlinx.coroutines.cancel
+
+/**
+ * Base class that provides a Kotlin/Native equivalent to the AndroidX `ViewModel`. In particular, this provides
+ * a [CoroutineScope][kotlinx.coroutines.CoroutineScope] which uses [Dispatchers.Main][kotlinx.coroutines.Dispatchers.Main]
+ * and can be tied into an arbitrary lifecycle by calling [clear] at the appropriate time.
+ */
+actual abstract class ViewModel {
+
+    actual val viewModelScope = MainScope()
+
+    /**
+     * Override this to do any cleanup immediately before the internal [CoroutineScope][kotlinx.coroutines.CoroutineScope]
+     * is cancelled in [clear]
+     */
+    protected actual open fun onCleared() {
+    }
+
+    /**
+     * Cancels the internal [CoroutineScope][kotlinx.coroutines.CoroutineScope]. After this is called, the ViewModel should
+     * no longer be used.
+     */
+    fun clear() {
+        onCleared()
+        viewModelScope.cancel()
+    }
+}
